+++ src/migrations/Farm.Migrations.PostgreSQL/Migrations/20260317024650_AddObicoServerApiKey.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Farm.Migrations.PostgreSQL.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddObicoServerApiKey : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "ApiKey",
+                table: "ObicoServers",
+                type: "character varying(500)",
+                maxLength: 500,
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "ApiKey",
+                table: "ObicoServers");
+        }
+    }
+}